feat: move feature in dedicated file
